[base] Introduce base::StrongAlias, use for JSDispatchHandle
Introduce a base::StrontAlias class which can wrap an integral type and
create a type-safe typedef that isn't a simple alias. This allows users
of a type to be explicit in whether they want to use the underlying
value, or pass around the type opaquely.

Use this for JSDispatchHandle.

Change-Id: I72cadb0e72ef1fae911b7683ac5f97c36426fe4f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6084776
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97663}