[heap] Fix DCHECK in Scavenger::ScavengeObject
When we find an object with a forwarding address, we update the slot and
check that the new target is valid. A valid target can be either in old
space or a large object (these are moved to old space only at the end of
the GC) or a pinned object. The DCHECK used to check if the target has a
self forwarding map word. This is not safe. We set forwarding addresses
with relaxed atomic and before the object is actually copied with the
assumption that other threads will only read the forwarding address but
never the object itself. Reading the map word to check if it is self
forwarding violated this assumption. As a result, the other thread saw a
map word of 0, the not yet initialized map word, and falsely considered
it as self forwarding.

Bug: 382988747
Change-Id: Ie1fb65bf7b097e813e37507ae8d136e95077c4fa
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6084780
Auto-Submit: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97660}