[sandbox] Validate instance type in HeapObject{Short}Print
Due to in-sandbox memory corruption, the object printing code may see
in-sandbox HeapObjects with the instance type of a TrustedObject. This
will then cause it to treat the object as a TrustedObject and for
example use the trusted space base when resolving compressed pointers
(if those are expected to point to other trusted objects). As these
pointers come from within the sandbox, we then quickly end up with
access violations in trusted space. This is probably all fine since for
printing, we're only ever reading from these objects, but we should
still fix this e.g. for fuzzer cleanliness.

Bug: 382928478
Change-Id: I60f3e2ac2eb75215f15f3dd2110b78f292369844
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6080992
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97661}