[regexp] Check the stack limit after PushCurrentPosition
These pushes were not accounted wrt RegExpStack::kStackLimitSlack. If
followed by a sequence of PushRegister, the latter could overflow.

This could be fixed in various ways, but (also since
PushCurrentPosition is called from 2 spots) this is the simplest.

Also, assert we're in bounds in PushRegister, which is now the only
method that does unchecked pushes to the backtracking stack.

Bug: 382244761
Change-Id: I273de7487bfcbd3456bd40ef325b642fea2899f4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6088690
Commit-Queue: Patrick Thier <pthier@chromium.org>
Auto-Submit: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97681}