[string] Iterative String::WriteToFlat
This adds a new iterative WriteToFlat implementation optimized
for the common case of strings constructed by repeatedly appending
to the end. Such strings are degenerate binary trees that are
unbalanced to the left. Our reverse-dfs iteration minimizes work
by always encountering the next relevant leaf naturally due to
iteration order.

We also optimize for the common case in which all leafs are
one-byte sequential strings.

Note this is a work in progress. The new implementation is added
as `WriteToFlat2` while we are still evaluating, and used only by
String::SlowFlatten.

Interesting:

- The dfs stack implementation intentionally uses functions instead
  of classes since classes cause clang to emit conservative code,
  missing a stack->register optimization that causes a 20% perf
  hit on arm64.
- The new WriteToFlat implementation is highly optimized for
  cons strings created through repeated appends. It's not clear yet
  if this is suitable to fully replace the existing WriteToFlat.
- The new implementation currently doesn't support flattening a
  substring of the given `src` string.

Drive-by: Clarify the (old) WriteToFlat signature, which is given a
`length`, not a `to` index.

Change-Id: Ibbd6da2cda49745009e954bba25c0708738dc93f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6062894
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97685}