[wasm][sandbox] Harden WasmTableObject::type() some more
We already had an SBXCHECK in place that the on-heap type had a
valid `kind()`. This CL takes that one step further to also check
that, if it's a reference type, its referenced index is in the
range that the associated module knows. The effect is that certain
kinds of in-sandbox corruption run into an SBXCHECK rather than
into a std::vector bounds check.

Fixed: 382326705
Change-Id: Ib3a71955e8d79b29189e97d58577e89f5e5f351f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6088060
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97730}