[wasm][fuzzer] Detect nondeterminism from memory.grow
Growing memory can nondeterministically fail, leading to false-positive
foozzie reports.
This CL checks the return value of the WasmMemoryGrow builtin and sets
the nondeterminism flag if -1 is returned. This also covers totally
deterministic failures when growing beyond the declared maximum size or
the platform-specific limit, but detecting this precisely would be a lot
of additional work to figure out the current size of the memory, the
passed delta, and the declared maximum.

Fixed: 380930509
Change-Id: I424d490cdc515c48d6bfbcd1e48f760916b999a1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6089074
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97716}