[context] Ensure heap number allocation before setting side data
When embedding heap numbers in optimized code and synchronizing via
acquire/release access to the side table property slot, it's
essential to allocate the heap number and set the context slot
value before updating the new slot property.

Failing to do so could lead to a state in the compiler thread where
the slot property is read as a mutable heap number and the slot value
is the previous value.

For consistency, we now always set the context slot value, even in
cases where the order of operations is not strictly necessary.

Bug: 383409592
Change-Id: Ic209cfe5efa0271c6cce5cc8e04c7c815a859844
Fixed: 382547590
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6087919
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97722}