[wasm] Use base::hasher for prefix hash computation
The base::hasher has a nice interface for adding more bytes to a running
hash. It also avoid using the very special "string hashing" that
GetWireBytesHash uses. Just adding the raw bytes should be more
efficient.

This will change the value of the prefix hash, but this is not
user-visible.

R=thibaudm@chromium.org

Bug: 383595408
Change-Id: I16916df4b7f08e4c4c75b19a00abb8bc4c46ff08
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6089956
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97749}