[foozzie] Fix ignored lines when output is capped
We now remove ignored lines in the raw bytes before output capping,
as such a line might be present in the non-crash run. The crashy run
however determines the byte length for comparisons and might cap the
to-be-ignored line right in the middle otherwise.

We also must continue doing the capping before decoding the
bytes, because of https://crbug.com/40914456.

No-Try: true
Fixed: 381384408
Change-Id: I29db0d808a89b3b9b20827dba13d233449c88ed0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6087812
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97771}