[tiering] Explicitly indicate what direction the budget should move
There are various reasons why we would want to update the budget.
Sometimes we simply want to reinitialize it. Sometimes we want to
delay interrupts. Sometimes we want to speed them up. So far we always
called the same API for all cases, which in some cases inadvertently
has the opposite effect if the budget is already further along. This
CL adds a flag to be explicit about where we want to go.

Change-Id: I44bed2774d2c4786d1fe2e4b2ef2545e1085f801
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6089298
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97817}