[wasm] Add skeleton of fuzz test for type canonicalization
This starts the implementation of a FuzzTest for type canonicalization.
Many features are still missing (recursion groups, final types,
inheritance), but those can be added later.

I have a prototype of recursion groups ready locally, but this requires
some changes to the WasmModuleBuilder. So we are landing this skeleton
now, also to check if the infrastructure works and the new fuzzer gets
picked up by ClusterFuzz.

R=jkummerow@chromium.org

Bug: 381687256
Change-Id: I483e80cae30b2f2289e9c9681ba54bf9f3cd81a5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6089297
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97795}