[turboshaft] Fix FrameState ID clashes with string escape analysis
In StringEscapeAnalyzer::ComputeFrameStatesToReconstruct, when a
FrameState has an elided StringConcat input, we mark the FrameStates
and all of its parents as "to reconstruct".

However, this is not enough, because when a FrameStates has an elided
StringConcat input, its children FrameStates should also be marked as
"to reconstruct". A possible fix would be to iterate FrameStates
forwards instead of backwards in ComputeFrameStatesToReconstruct and
check parent FrameState "to reconstruct" status (cf draft
implementation https://crrev.com/c/6098278). However, this wouldn't be
enough, because of cases like this:

    v1: FrameState(...)
    v2: FrameState(v1, ...)
    v3: FrameState(v1, StringConcat, ...)

Here, hen visiting `v1` and `v2`, we'll decide to not reconstruct, but
when visiting `v3`, we'll decide to reconstruct `v1`. So, we'll end up
reconstructing `v1` and not `v2`. And this could be wrong, because
reconstructing FrameState changes IDs, which means that `v2` could end
up using wrong IDs.

One way to fix this would be to use union-find sets: in the example
above, `v2` would be in the same set as `v1` because it's a child of
`v1`, and then `v3` would also be in the same set as `v1` for the same
reason, and would thus also be in the same set as `v2`. And when
visiting `v3`, we would mark the whole set as "to
reconstruct". However, this is a bit more complex than what the
current CL does, and it would lead to needlessly reconstruct more
FrameStates than is really needed.

So, the fix in the current CL: we keep separate ID sets for
dematerialized objects and dematerialized strings, and unify them in
the instruction selector. This should be seen as a temporary state:
long-term, we should do a single escape analysis pass, which creates a
single kind of IDs.

Fixed: chromium:382547699
Change-Id: I570d987bd355cc079d6c8e39d80f3c9f913150a1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6097638
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97851}