[regexp] Parser: Fix position() at surrogate pair
position() when current() is a surrogate pair returned the wrong
position (in the middle of the surrogate pair).
This is a problem when we save and later restore the current position,
as we skip the lead-surrogate when we continue parsing after the reset.

Luckily, the only case where this actually lead to a problem is rather
niche: If a backreference is directly followed by a surrogate pair and
we haven't parsed a matching capture group yet.

Fixed: 384605103
Change-Id: Ib4e794fe6a82bdbb36a703fb4950cb856df7c9fb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6097646
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Auto-Submit: Patrick Thier <pthier@chromium.org>
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97852}