[roots] Remove ReadOnlyRoots Handle getters
ReadOnlyRoots can be initialized with either an Isolate/Heap, or a
ReadOnlyHeap. These have separate tables of roots; one is the
per-Isolate RootsTable, the other is an array of values in the
ReadOnlyHeap object which is a copy/cache of the root values.

If ReadOnlyRoots is initialized with a ReadOnlyHeap (e.g. with the
GetReadOnlyRoots() getter), then handles created by it won't be
legitimate root handles, but will be handles into this ReadOnlyHeap
cache. This would mean that they fail handle identity checks like
IsRootHandle.

Instead, we should only allow getting Tagged values from ReadOnlyRoots,
and anything that wants a ReadOnlyRoot Handle should read it via the
roots table (in practice, this means via the existing accessors on
Factory).

Change-Id: I6322a1c7baf3bbc0dc159b984467874b436bf379
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6106570
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97862}