Change WriteShortData's function signature to enforce uint8_t on input.
Right now WriteShortData takes a uintptr_t as input and statically casts that input to a uint8_t without checking that it's a valid size for that cast.

Currently the call site for WriteShortData checks this via a DCHECK, but it would be clearer and prevent future callers from mishandling this by having WriteShortData take a uint8_t in the first place.

This was found via a CodeQL query:
https://chromium-review.googlesource.com/c/chromium/src/+/6020237

Based on a previous security bug whose root cause involved type confusion after a static cast:
https://g-issues.chromium.org/u/0/issues/351327767

Change-Id: I7c600d020db99b0197dd2d6d78eac67369aef232
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6040738
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Julia Hansbrough <flowerhack@google.com>
Cr-Commit-Position: refs/heads/main@{#97973}