[wasm] Forbid allocations while "thread in wasm" is set
In the end it was just two runtime functions that were called from Wasm
and were missing a `SaveAndClearThreadInWasmFlag` scope.

The most tricky part was that calling `IsTrapHandlerEnabled()` disallows
enabling the trap handler later, and we sometimes have allocations
before enabling the trap handler. This CL thus introduces a new
`AssertThreadNotInWasm` function which checks the equivalent of
`!IsTrapHandlerEnabled() || !IsThreadInWasm()` without that side effect.

We now have more TODOs to get rid of the "thread in wasm" flag, so we
might want to prioritize that soon.

Bug: 384605106, 40192807

R=jkummerow@chromium.org
CC=mseaborn@chromium.org

Change-Id: I2748b09e99130607c0e77ddc8774e047d5ded3d3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6097964
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97888}