[wasm] Use C++ allocated buffer in generic js-to-wasm wrapper
The generic js-to-wasm wrapper can be used for calls from JavaScript to
any WebAssembly function. It does that by interpreting the signature of
the WebAssembly function, and by converting all incoming parameters
according to that signature. The converted parameter then get stored in
a byte buffer, which gets loaded into the correct registers and stack
slots in a small assembly snippet.

A WebAssembly function is allowed to have 1000 parameters, and the byte
buffer has to have enough space for that case. A byte buffer for 1000
parameters is, however, too big to be allocated on the stack. Therefore
a ByteArray was allocated on the heap so far if the number of parameters
exceeded 10. The ByteArray, however, can be used for a V8 heap sandbox
escape. With this CL, a byte buffer gets allocated and stored in the
IsolateData. This byte buffer gets loaded through the root register and
is therefore not accessible through memory corruption in the V8 heap.

The byte buffer gets allocated lazily when generic js-to-wasm wrapper
gets used for the first time, so that no startup overhead gets
introduced when no WebAssembly is used.

Bug: 384844209

Change-Id: Ia4f5d866e8a709d27a471776e3e1b28ba661593e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6108765
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97903}