[maglev] Ensure we don't attempt osr if it's blocked
If OSR to Turbofan is blocked for any reason we must return 0 in
CompileOptimizedOSRFromMaglev. Otherwise TryOnStackReplacement in maglev
will deoptimize in preparation for an OSR to Turbofan, which will not
happen. We'll end up in the same Maglev OSR code and fail again and
again.

Bug: 385775385

Change-Id: I42d5b8bd4e5ebfecb601d7158a8bbd22f0ca2131
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6110220
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98071}