[shadow stack] Lazy deoptimization of baseline code
Sparkplug code can be also deoptimized for debug purpose. V8 currently
updates return addresses directly to discard baseline code on stack,
which violates CET shadow stack.

This CL takes a similar approach as lazy deoptimization of maglev or
turbofan. Reserve 5 bytes after each call as NOP in the compiled code,
and patch the reserved nop to a jump to deoptimization instead of
updating the return addresses on stack directly.

Since sparkplug deoptimization does not have the same deopt infra as
higher tier compilers, and directly patches the bytecode offset after
reframing, the frame should be detected as InterpretedFrame once the
code is patched for deoptimization. So, we mark the baseline code for
deoptimization when we patch the baseline code, and interpret the new
stack frame as InterpretedFrame if the baseline code was marked for
deoptimization.

Bug: 42201233
Change-Id: I30e5ad2b6270e1adad456f6acf90e81d645ed783
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6076765
Commit-Queue: Choongwoo Han <choongwoo.han@microsoft.com>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97943}