[js_fuzzer] Refactoring - move flag-loading logic into corpus
This enables creating corpus sub classes with logic specific to a
particular corpus. We specialize a V8 corpus and move the V8-specific
flag-loading logic to it. In test cases from other places, this logic
should not be applied (this resolves an old TODO).

This also prepares creating a Fuzzilli corpus with its own flag logic
in a follow up.

Some test data is moved, since a V8 corpus wants its files in a v8
base directory.

Bug: 362963886
Change-Id: Ie77618ef96dde5040ef156aed387f157d94d731a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6148431
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#97963}