Mark old String::Write* APIs as deprecated
As part of https://crbug.com/373485796 we introduced new versions of
these APIs. With this CL, we're now marking the old APIs as DEPRECATED.
Some time in the future, we will first delete these APIs, then rename
the new APIs to the old names (so drop the V2 suffix).

Bug: 373485796
Change-Id: I1df1b676315246f193df08dd8b9d71ed2ea3beef
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6148579
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98013}