[sandbox] Use fine-grained type checks for D8 external objects
Prior to this CL we were using the same ExternalPointerTag (the
kGenericManagedTag) for different objects used by d8. This meant that it
was possible to cause type confusions between these objects, and fuzzers
eventually figured out how to do that. Now that we have more tags
available after https://crrev.com/c/6055043 this is easy to fix by using
dedicated tags for the objects.

Bug: 385406656
Change-Id: Ia2de5a68da1df2d3883fb4c0893c6e4a7fd02f55
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6157475
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97998}