[wasm] Handle WKI assumptions during background compilation
For some reason we did only handle well-known imports assumptions on the
single-function `AddCompiledCode` method which is only used in
exceptional cases (including the `WasmTierUpFunction` runtime function
used in tests).

This CL refactors the API of `AddCompiledCode` to return a struct (or
multiple in case we add code for multiple function) that holds both the
`WasmCode` pointer and a pointer to the assumptions. This is then passed
on to the "publish" phase, where the assumptions are checked
consistently.

In order to trigger the bug more reliable this also adds a
`WasmTriggerTierUpForTesting` runtime function which triggers tier-up in
the background (like in production).

R=jkummerow@chromium.org

Fixed: 382326704
Change-Id: I8093cf4f4838dfe80e2169887f4ddce89c5fbe39
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6109527
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#97983}