[wasm][sandbox] cache signature hash in the CanonicalSig
to avoid recalculating it all the time.
In the cases where we need non-canonical signature hashes, we look
them up through the type canonicalizer.

Plus, the signature hash calculation in the CallDescriptor will now only
happen for indirect call descriptors. It might be possible to avoid
this by passing the hash through in more places, but it would require
quite a lot of refactoring.

Change-Id: I6515a361b40820aa6a537da68ecf82c81ebbf3f6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6094047
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#97987}