[runtime] Fix ArrayBuffer creation attaching extension objects
ArrayBufferExtension objects are witness objects to track BackingStore
objects. The are attached to JSABs in a 1:1 fashion. The initial setup
through builtins is convoluted as AB creation is observable and should
preceed the actual backing store creation. Since the backing store
creation can fail this leads to a half initialized temporary state
where an AB is present but no backing store is attached and thus no
extension objects is present.

The current code works on the assumption of a well-behaving program.

In a world where we follow sandbox rules malloced state is trusted and
always has to be verified when being read from the heap. In order to
avoid verification, the code is refactored to only a have a single
setup phase. This allows us to work with the extension when it's
initially allocated and prevents the mutator from reusing the object
and tinkering with the lists it is contained in.

Change-Id: I9d2f23f3ad141aea8b1543ac0d071ecc688e173b
Bug: 384186547
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6149097
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98001}