[sandbox] Only use bits [48, 56) for EPT pointer tagging
The top byte may be used for other purposes, for example HWASan or MTE.
As such, with this CL we're now moving the type tag and marking bit into
the 2nd most significant byte of the pointer and leave the top byte
completely untouched. On Arm64, this requires the same number of
instructions (we now use UBFX). On x64 it requires additional
instructions, but the overall cost should still be dominated by the
memory load, not the (constant-time) untagging instructions.

Bug: 381444303, 372157557, 388376335
Change-Id: Iaaae388b48a6c050962641955921cd2038c74ec7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6162735
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98027}