[turbofan] Handle missing source positions when profiling is enabled
In some cases we might fail to collect source positions when profiling
is enabled. One case where this can happen is when we run out of stack
space and we cannot reparse. This CL makes sure that TF doesn't assume
that source positions are available when this happens.

Bug: 381390367
Change-Id: I4b5d95ac09a31d037df924bcf3c5280fd63b8853
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6163444
Auto-Submit: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98029}