[wasm] Add check in else case of type switch
The generic js-to-wasm wrappers and wasm-to-js wrappers do not support
invalid signatures. If a signature is invalid, a wrapper is used that
throws an exception unconditionally.

Since signatures are therefore valid by design, the generic wrappers
only checked with DCHECKs that all types in the incoming signature are
valid. However, with corruption within the heap sandbox, it is still
possible that the generic wrapper has to handle invalid types.

This CL therefore changes the DCHECKs into CHECKs, so that if an
attacker manages to call the generic wrappers with an invalid signature,
V8 crashes and does not continue execution with a type confusion.

R=clemensb@chromium.org

Bug: 385038763
Change-Id: I82bd1b52b34019063583b75fe0d6f7018f959d60
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6163545
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98036}