[sandbox] Harden the DoubleToStringWithRadix runtime function
The way this helper function is currently implemented means that it will
corrupt stack memory (essentially a wild-write) when given invalid
double values (e.g. NaN or Infinity). Normally this cannot happen
because callers are responsible for checking for these values first.
However, in the sandbox attacker model, this is tricky to get right as a
Number's value can change in between due to heap corruption. As such,
this CL now hardens the runtime function to terminate the process with a
controlled CHECK failure instead of corrupting memory.

Bug: 388616182
Change-Id: I18eedfbc74ceef375741698b86a9a605a094010e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6165025
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98050}