[fuzzer][wasm] Generate shift of local values
Generates patterns like:
  local.set 0 (local.get 1)
  local.set 1 (local.get 2)
  local.set 2 (local.get 3)
  local.get 0

When this is emitted in a loop, the value of local 3 slowly progresses
towards local 0, needing multiple iterations to "reach" it.
This is an interesting pattern for typed optimizations:
If the first 3 locals contain a structref and the 4th an anyref, these
are the refined types at the beginning of the loop on each iteration:
0:  [structref, structref, structref, anyref]
1:  [structref, structref, anyref,    anyref]
2:  [structref, anyref,    anyref,    anyref]
3+: [anyref,    anyref,    anyref,    anyref]

This change also changes the implementation of get_local_ref() which
previously only picked a random local and would just fail if the local
wasn't of the requested type.
Instead it now picks a random local and if it doesn't have a matching
type, it iterates over the locals starting from the random local trying
to find a local with a matching type.
This increases the chance of being able to emit a local.get.

Bug: 383356864, 383740723
Change-Id: Ic98f69a2a4baa915b098394484588d2c730719cc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6163632
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98055}