[heap] Skip checking full addresses in CSS
For each 64bit address found on the stack during CSS, we first try to
match the full 64bit address to an object, then treat each of the 32bit
halves of the address as a compressed pointer, decompress them and try
to match them to an object.
If the 64bit address points to within a cage, the decompressed lower
32bit half of it would point to the same address, and we will be
handling the address twice.

Optimize by skipping the full 64bit address in builds with pointer
compression enabled.

Bug: 379788114
Change-Id: I5b14e6e2edaba93ba4e058e8b04150a4734314aa
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6167109
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98057}