[heap] Bump thread QoS class for safepointing
Bump the quality-of-service class on Darwin when safepointing running
threads. Without this we end up with a priority-inversion where the
high-priority main thread blocks on lower priority background threads.

This is necessary because for safepointing there is no explicit
mechanism to automatically transfer the boost to the background
threads. Background threads need to run any random/regular code until
they reach a safepoint. The kernel cannot othwerise know about this
dependency.

This might cause the renderer hangs in EnterLocalSafepointScope(),
which seems to mostly happen on Darwin.

Bug: 369436598
Change-Id: I4ef3f76f97d7ae69e19c243d3862c2b62dbc484e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6165198
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98084}