[api] Fix micro task queue completion callbacks
We used the non-empty state of the COW vector to indicate whether the
COW vector was used in first place. This is not correct if the COW
vector is used and all callbacks (i.e. the last one) was removed from
within the callback itself.

Fix this by indicating this bit via using optional.

Bug: 350350588
Change-Id: I03c629ce72d25d854f28e79e7e5e435052585230
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6172299
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98090}