[wasm] Use dynamically-allocated buffer in generic js-to-wasm wrapper
The generic js-to-wasm wrapper can be used for calls from JavaScript to
any WebAssembly function. It does that by interpreting the signature of
the WebAssembly function, and by converting all incoming parameters
according to that signature. The converted parameter then get stored in
a byte buffer, which gets loaded into the correct registers and stack
slots in a small assembly snippet.

A WebAssembly function is allowed to have 1000 parameters, and the byte
buffer has to have enough space for that case. A byte buffer for 1000
parameters is, however, too big to be allocated on the stack. For
signatures with more than 10 parameters a ByteArray could therefore be
allocated on the V8 heap. However, a ByteArray on the V8 heap could be
used for a sandbox escape. Alternatively a byte array could be stored in
the isolate, and then used by the generic wrapper. However, multiple
activations of the generic wrapper could exist at the same time, so
a single buffer would be insufficient*.

With this CL, a buffer gets allocated dynamically in C++, and
deallocated before the generic wrapper returns. For that the
implementation of the generic wrapper gets wrapped into a try-catch, to
make sure the buffer also gets freed when an exception happened. After
deallocating the buffer, an exception gets re-thrown.

* A simple recursion that alternates between JS and wasm would not be a
  problem, a single buffer would be sufficient for that. However, the
  conversion of each argument could trigger the execution of arbitrary
  JS, including calls to wasm, and for this scenario a single buffer is
  insufficient.

R=clemensb@chromium.org
CC=thibaudm@chromium.org

Bug: 385256110
Change-Id: I8566e255cc6e1799036a3febccae0184c413bca1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6171272
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98097}