[13.3][wasm] Use dynamically-allocated buffer in generic js-to-wasm wrapper
The generic js-to-wasm wrapper can be used for calls from JavaScript to
any WebAssembly function. It does that by interpreting the signature of
the WebAssembly function, and by converting all incoming parameters
according to that signature. The converted parameter then get stored in
a byte buffer, which gets loaded into the correct registers and stack
slots in a small assembly snippet.

A WebAssembly function is allowed to have 1000 parameters, and the byte
buffer has to have enough space for that case. A byte buffer for 1000
parameters is, however, too big to be allocated on the stack. For
signatures with more than 10 parameters a ByteArray could therefore be
allocated on the V8 heap. However, a ByteArray on the V8 heap could be
used for a sandbox escape. Alternatively a byte array could be stored in
the isolate, and then used by the generic wrapper. However, multiple
activations of the generic wrapper could exist at the same time, so
a single buffer would be insufficient*.

With this CL, a buffer gets allocated dynamically in C++, and
deallocated before the generic wrapper returns. For that the
implementation of the generic wrapper gets wrapped into a try-catch, to
make sure the buffer also gets freed when an exception happened. After
deallocating the buffer, an exception gets re-thrown.

* A simple recursion that alternates between JS and wasm would not be a
  problem, a single buffer would be sufficient for that. However, the
  conversion of each argument could trigger the execution of arbitrary
  JS, including calls to wasm, and for this scenario a single buffer is
  insufficient.

R=clemensb@chromium.org
CC=thibaudm@chromium.org

Bug: 385256110

(cherry picked from commit 14cbf9bdbb97057a9c1f609ace381bd086fcfd56)

Change-Id: I5e5984b606acc8b8e1f3556954d47ce956a2533f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6182060
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/branch-heads/13.3@{#20}
Cr-Branched-From: 41dacffe436aeb9311879cb07648f1e36609a804-refs/heads/13.3.415@{#1}
Cr-Branched-From: 3348638c0af67c885b30891a358c89a917ac9759-refs/heads/main@{#97937}