[runtime] Clear thread-in-wasm flag in StringToNumber
TurboFan's StringToNumber node is not part of the control or effect
chain, so it can move freely.
After inlining the js-to-wasm wrapper, it can therefore move within the
section where we set the "thread in wasm" flag. Since StringToNumber can
allocate, we need to clear that bit during the runtime function.

This issue indicates that it's becoming increasingly difficult to define
when "a thread is executing Wasm". A better long-term solution would be
to get rid of that flag completely (issue 40192807).

R=jkummerow@chromium.org

Bug: 387055479, 384605106, 40192807
Change-Id: I90fc64cb06b924a1c6e5c664817dcc3e285f80a9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6172628
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98104}