[runtime][torque] Undefined Doubles
Introduce a special nan value to represent undefined values in
doubles to allow JSArrays to stay in (holey) double elements instead
of transitioning into tagged representation.

This is only the first part of a larger change as proper handling in
the compiler is still missing.

Bug: 385155404
Change-Id: I1c1c9418447d0f060dca5b28a63b858b4e1a6dd5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6063479
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98120}