[wasm][test] Enable status file checks
I randomly found those status file checks, and we can easily enable them
for the wasm-js and wasm-spec-tests suites. This finds dozens of status
file entries for tests that do not exist any more, either because they
were removed, or because of rebases such that the file in the proposal
repo is identical to the upstream version, or unchanged since the
branch. We do not run those tests since https://crrev.com/c/5920423.

R=machenbach@chromium.org, jkummerow@chromium.org

Bug: 384676197
Change-Id: I02925cc3763e3dabe6bcc29d1db6ea628691db12
Fixed: 364913810, 378077852
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6172286
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98134}