[etw] Fix ETW events on Capture-State for the v8 ETW provider
A recent commit (
    https://chromium-review.googlesource.com/c/v8/v8/+/5918068)
implemented capture state support for the v8 ETW provider.

This allows asking the v8 EventSource to emit events for all jitted
functions in the rundown phase (at the end of a tracing session), which
is very useful when using a circular buffer for the trace. (See
https://learn.microsoft.com/en-us/windows-hardware/test/wpt/capturestateonsave
for more details).

However, there was a problem in the original implementation because
SourceLoad and MethodLoad events emitted during a Capture-State requests
need to specify a different events (Source_DCStart, Method_DCStart) to
be correctly handled by Windows.
This CL fixes the implementation accordingly.

Bug: 1346064442
Change-Id: I972ea451dca33c77507463dde8c6c81adfaed907
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6093311
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Camillo Bruni <cbruni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98155}