[wasm] Fix recognize-imports test for multi-isolate execution
Remove an assertion which does not always hold if the test executes in
parallel in multiple isolates.

The alternative would be to skip the test in that configuration, but
having test coverage for that configuration is more important than that
one assertion in the test.

R=jkummerow@chromium.org

Fixed: 388869361
Change-Id: Ie95c9a44c9f59d27cdcb896ec1f4100ddda443a8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6175162
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98127}