[wasm] Fixes for 32-bit WasmCodePointer
A WasmCodePointer is always 32-bit after https://crrev.com/c/6062889.
This CL updates the call to CallIndirectIC in Liftoff, which always
treated the code pointer as an intptr-sized value (which was already
inconsistent before that CL).

We also change a `shlq` to `shll` on x64, which saves the `REX.W` byte
in generated code.

R=jkummerow@chromium.org

Bug: 389879224
Change-Id: Id649225ba37c67038f4886af80c01835446fd1da
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6175349
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#98171}